+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Connections/Controller/Internal/DeleteConnectionAction.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection;
+namespace Akeneo\Connectivity\Connection\Infrastructure\Connections\Controller\Internal;
 
-use Akeneo\Connectivity\Connection\Application\Settings\Command\RegenerateConnectionPasswordCommand;
-use Akeneo\Connectivity\Connection\Application\Settings\Command\RegenerateConnectionPasswordHandler;
+use Akeneo\Connectivity\Connection\Application\Settings\Command\DeleteConnectionCommand;
+use Akeneo\Connectivity\Connection\Application\Settings\Command\DeleteConnectionHandler;
 use Oro\Bundle\SecurityBundle\SecurityFacade;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -17,10 +17,10 @@ use Symfony\Component\Security\Core\Exception\AccessDeniedException;
  * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-final class RegeneratePasswordAction
+final class DeleteConnectionAction
 {
     public function __construct(
-        private RegenerateConnectionPasswordHandler $regenerateConnectionPasswordHandler,
+        private DeleteConnectionHandler $deleteConnectionHandler,
         private SecurityFacade $securityFacade,
     ) {
     }
@@ -31,13 +31,13 @@ final class RegeneratePasswordAction
             throw new AccessDeniedException();
         }
 
-        $command = new RegenerateConnectionPasswordCommand($request->get('code', ''));
+        $command = new DeleteConnectionCommand($request->get('code', ''));
         try {
-            $password = $this->regenerateConnectionPasswordHandler->handle($command);
+            $this->deleteConnectionHandler->handle($command);
         } catch (\Exception $e) {
             return new JsonResponse(['message' => $e->getMessage()], Response::HTTP_BAD_REQUEST);
         }
 
-        return new JsonResponse(['password' => $password]);
+        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
     }
 }
