+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/routing.yml
@@ -16,7 +16,6 @@ akeneo_connectivity_connection_connect_apps_v1_openid_public_key:
     methods: [ GET ]
 
 # Internal API
-
 akeneo_connectivity_connection_error_management_rest:
     resource: ./ErrorManagement/routing.yml
 
@@ -24,10 +23,7 @@ akeneo_connectivity_connection_audit_rest:
     resource: ./Audit/routing.yml
 
 akeneo_connectivity_connection_rest:
-    resource: ./routing/connection.yml
-    name_prefix: akeneo_connectivity_connection_rest_
-    prefix: /rest/connections
-    trailing_slash_on_root: false
+    resource: ./Connections/routing.yml
 
 akeneo_connectivity_connection_settings_rest:
     resource: ./Settings/routing.yml
