--- src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/routing/connection.yml
@@ -1,49 +1,0 @@
-list:
-    path: ''
-    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\ListConnectionsAction
-    methods: [GET]
-
-create:
-    path: ''
-    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\CreateConnectionAction
-    methods: [POST]
-
-get:
-    path: '/{code}'
-    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\GetConnectionAction
-    methods: [GET]
-    requirements:
-        code: '[a-zA-Z0-9_]+'
-
-update:
-    path: '/{code}'
-    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\UpdateConnectionAction
-    methods: [POST]
-    requirements:
-        code: '[a-zA-Z0-9_]+'
-
-delete:
-    path: '/{code}'
-    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\DeleteConnectionAction
-    methods: [DELETE]
-    requirements:
-        code: '[a-zA-Z0-9_]+'
-
-regenerate_secret:
-    path: '/{code}/regenerate-secret'
-    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\RegenerateSecretAction
-    methods: [POST]
-    requirements:
-        code: '[a-zA-Z0-9_]+'
-
-regenerate_password:
-    path: '/{code}/regenerate-password'
-    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Connection\RegeneratePasswordAction
-    methods: [POST]
-    requirements:
-        code: '[a-zA-Z0-9_]+'
-
-wrong_credentials_combination_list:
-    path: '/wrong-credentials-combination'
-    controller: Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\WrongCredentialsCombination\ListWrongCredentialsCombinationsAction
-    methods: [GET]
