--- src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/wrong_credentials_connection.yml
@@ -1,5 +1,0 @@
-services:
-    Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\WrongCredentialsCombination\ListWrongCredentialsCombinationsAction:
-        public: true
-        arguments:
-            - '@Akeneo\Connectivity\Connection\Infrastructure\Connections\WrongCredentialsCombination\Persistence\DbalWrongCredentialsCombinationRepository'
