+++ src/Akeneo/Connectivity/Connection/back/tests/EndToEnd/Connection/Controller/Internal/RegenerateConnectionPasswordEndToEnd.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Connectivity\Connection\Tests\EndToEnd\Connection\Controller\Internal;
+
+use Akeneo\Connectivity\Connection\back\tests\EndToEnd\WebTestCase;
+use Akeneo\Connectivity\Connection\Domain\Settings\Model\ValueObject\FlowType;
+use Akeneo\Test\Integration\Configuration;
+use PHPUnit\Framework\Assert;
+use Symfony\Component\HttpFoundation\Response;
+
+/**
+ * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
+ * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ */
+class RegenerateConnectionPasswordEndToEnd extends WebTestCase
+{
+    public function test_it_regenerates_a_connection_password(): void
+    {
+        $connection = $this->createConnection('franklin', 'Franklin', FlowType::DATA_SOURCE, false);
+
+        $this->authenticateAsAdmin();
+        $this->client->request('POST', '/rest/connections/franklin/regenerate-password');
+        $result = \json_decode($this->client->getResponse()->getContent(), true);
+
+        Assert::assertEquals(Response::HTTP_OK, $this->client->getResponse()->getStatusCode());
+        Assert::assertArrayHasKey('password', $result);
+        Assert::assertNotSame($connection->password(), $result['password']);
+    }
+
+    protected function getConfiguration(): Configuration
+    {
+        return $this->catalog->useMinimalCatalog();
+    }
+}
