+++ src/Akeneo/Connectivity/Connection/back/tests/EndToEnd/Connection/Controller/Internal/RegenerateConnectionSecretEndToEnd.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\back\tests\EndToEnd\Connection;
+namespace Akeneo\Connectivity\Connection\Tests\EndToEnd\Connection\Controller\Internal;
 
 use Akeneo\Connectivity\Connection\back\tests\EndToEnd\WebTestCase;
 use Akeneo\Connectivity\Connection\Domain\Settings\Model\ValueObject\FlowType;
@@ -14,19 +14,20 @@ use Symfony\Component\HttpFoundation\Response;
  * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class RegenerateConnectionPasswordEndToEnd extends WebTestCase
+class RegenerateConnectionSecretEndToEnd extends WebTestCase
 {
-    public function test_it_regenerates_a_connection_password(): void
+    public function test_it_regenerates_a_connection_secret(): void
     {
-        $connection = $this->createConnection('franklin', 'Franklin', FlowType::DATA_SOURCE, false);
+        $this->createConnection('franklin', 'Franklin', FlowType::DATA_SOURCE, false);
 
         $this->authenticateAsAdmin();
-        $this->client->request('POST', '/rest/connections/franklin/regenerate-password');
+        $this->client->request('POST', '/rest/connections/franklin/regenerate-secret');
         $result = \json_decode($this->client->getResponse()->getContent(), true);
 
-        Assert::assertEquals(Response::HTTP_OK, $this->client->getResponse()->getStatusCode());
-        Assert::assertArrayHasKey('password', $result);
-        Assert::assertNotSame($connection->password(), $result['password']);
+        $expectedResult = null;
+
+        Assert::assertEquals(Response::HTTP_NO_CONTENT, $this->client->getResponse()->getStatusCode());
+        Assert::assertEquals($expectedResult, $result);
     }
 
     protected function getConfiguration(): Configuration
